requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
pyyaml>=6.0
watchdog>=2.1.0
kagglehub>=0.2.0
matplotlib>=3.5.0
pillow>=9.0.0
flask>=2.0.0
scikit-learn>=0.24.0